// web/src/generated/models/metadataEnumValueUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */

export interface MetadataEnumValueUpdate {
  label?: string | null;
  description?: string | null;
}